src/MeteorRush/settings.py
import os

from dotenv import find_dotenv, load_dotenv

load_dotenv(find_dotenv())

DB_SERVER = os.getenv("DB_SERVER")
DB_DATABASE = os.getenv("DB_DATABASE")
DB_USER = os.getenv("DB_USER")
DB_PASSWORD = os.getenv("DB_PASSWORD")
